+++ entry/src/main/ets/pojo/index.ts
@@ -1,0 +1,9 @@
+export type FileUploadResp = {
+  downloadSignUrl?: string
+  fileKey?: string
+  urlMd5?: string
+  signUrl?: string
+  host?: string
+  channelCode?: string
+  contentType?: string
+}